requirements.txt
dnspython
asyncio
colorama
optparse